Refactor out state delta handling into its own class (#4917)
